+++ internal/cli/move.go
@@ -5,7 +5,6 @@ import (
 	"log/slog"
 	"os"
 	"path/filepath"
-	"syscall"
 
 	cp "github.com/otiai10/copy"
 )
@@ -29,33 +28,6 @@ func copyAndDelete(src, dst string) error {
 	}
 
 	return nil
-}
-
-// isSamePartition checks if the source and destination reside on the same filesystem partition.
-func isSamePartition(src, dst string) (bool, error) {
-	srcStat, err := os.Stat(src)
-	if err != nil {
-		return false, fmt.Errorf("failed to get source file stats: %w", err)
-	}
-
-	dstStat, err := os.Stat(dst)
-	if err != nil {
-		return false, fmt.Errorf("failed to get destination file stats: %w", err)
-	}
-
-	srcSys := srcStat.Sys().(*syscall.Stat_t)
-	dstSys := dstStat.Sys().(*syscall.Stat_t)
-
-	// Compare the device identifiers (st_dev) of the source and destination
-	// If the device IDs are the same, the files are on the same partition.
-	samePartition := srcSys.Dev == dstSys.Dev
-
-	slog.Debug("check src/dst file info",
-		"samePartition", samePartition,
-		"src st_dev", srcSys.Dev,
-		"dst st_dev", dstSys.Dev)
-
-	return samePartition, nil
 }
 
 // move attempts to rename a file or directory. If it's on different partitions, it falls back to copying and deleting.
